rtc/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
 
ultralytics>=8.0.0

transformers>=4.35.0
accelerate>=0.24.0

opencv-python>=4.8.0
Pillow>=10.0.0

numpy>=1.24.0
opencv-python

fastapi
uvicorn
pydantic